+++ samples/AuthRemoteIdentity/AuthRemoteIdentityAppHost/Program.cs
@@ -1,4 +1,3 @@
-using Microsoft.AspNetCore.SystemWebAdapters.Aspire;
 using System.Security.Policy;
 using Microsoft.Extensions.Configuration;
