+++ samples/ServiceDefaults/Extensions.cs
@@ -141,18 +141,5 @@ public static class SampleServiceExtensions
 
         return app;
     }
-
-    public static IEndpointConventionBuilder MapRemoteAppFallback(this WebApplication app, [StringSyntax("Route")] string? pattern = "/{**catch-all}")
-    {
-        var url = app.Services.GetRequiredService<IOptions<RemoteAppClientOptions>>().Value.RemoteAppUrl.OriginalString;
-
-        return app.MapForwarder(pattern, url)
-
-            // If there is a route locally, we want to ensure that is used by default, but otherwise we'll forward
-            .WithOrder(int.MaxValue)
-
-            // If we're going to forward the request, there is no need to run any of the middleware after routing
-            .ShortCircuit();
-    }
 #endif
 }
