+++ samples/SessionRemote/SessionRemoteAppHost/SessionRemoteAppHost.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Microsoft.AspNetCore.SystemWebAdapters.Aspire\Microsoft.AspNetCore.SystemWebAdapters.Aspire.csproj" IsAspireProjectResource="false" />
+    <ProjectReference Include="..\..\..\src\Aspire.Hosting.IncrementalMigration\Aspire.Hosting.IncrementalMigration.csproj" IsAspireProjectResource="false" />
     <ProjectReference Include="..\SessionRemoteCore\SessionRemoteCore.csproj" />
     <ProjectReference Include="..\SessionRemoteFramework\SessionRemoteFramework.csproj" />
   </ItemGroup>
