+++ samples/SessionRemote/SessionRemoteCore/Program.cs
@@ -5,9 +5,6 @@ using Microsoft.AspNetCore.SystemWebAdapters.Features;
 var builder = WebApplication.CreateBuilder(args);
 
 builder.AddServiceDefaults();
-
-builder.Services.AddReverseProxy();
-
 builder.AddSystemWebAdapters()
     .AddSessionSerializer(options =>
     {
