+++ src/Aspire.Hosting.IncrementalMigration/AspireConstants.Shared.cs
