+++ src/Aspire.Microsoft.AspNetCore.SystemWebAdapters/AspireConstants.cs
