--- src/Microsoft.AspNetCore.SystemWebAdapters.FrameworkServices/AspireConstants.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.AspNetCore.SystemWebAdapters;
-
-internal static partial class AspireConstants
-{
-    const string Separator = ":";
-}
