+++ src/Microsoft.AspNetCore.SystemWebAdapters.FrameworkServices/Microsoft.AspNetCore.SystemWebAdapters.FrameworkServices.csproj
@@ -15,9 +15,6 @@
     <None Include="Content\*" Pack="true" PackagePath="Content\" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Microsoft.AspNetCore.SystemWebAdapters.Aspire\AspireConstants.Shared.cs" Link="AspireConstants.Shared.cs" />
-  </ItemGroup>
-  <ItemGroup>
     <Reference Include="System.Web" />
     <Reference Include="System.Net.Http" />
     <Reference Include="System.Configuration" />
